教师和学生批量导入 教师和学上模板下载 教师和管理员修改密码 作业项目增加参考方案文件(增删改)

+++ src/main/java/com/hit/curricelumdesign/service/WorkProjectService.java
@@ -68,6 +68,11 @@ public class WorkProjectService {
         if (CollectionUtils.isEmpty(fileList) || param.getFileIdList().size() != fileList.size()) {
             throw new BaseException(Error.WORK_PROJECT_ASSOCIATE_FILE_HAS_NOT_EXIST);
         }
+        //2020-01-05增加参考方案的部分，这里参考方案是否可以为空
+        List<File> referenceSolutionList = fileManager.getFileByIdList(param.getReferenceSolutionList());
+        if (CollectionUtils.isEmpty(referenceSolutionList) || param.getReferenceSolutionList().size() != referenceSolutionList.size()) {
+            throw new BaseException(Error.WORK_PROJECT_ASSOCIATE_FILE_HAS_NOT_EXIST);
+        }
         Date now = new Date();
         WorkProject workProject = new WorkProject();
         BeanUtil.copyProperties(param, workProject, "fileIdList");
@@ -81,12 +86,28 @@ public class WorkProjectService {
         List<WorkProjectFile> workProjectFileList = new ArrayList<>();
         for (File file : fileList) {
             WorkProjectFile workProjectFile = buildWorkProjectFile(file, workProject);
+            //设置文件的使用类型为资料
+            workProjectFile.setFileType(Constants.File.FileTypes.INFORMATION.getStatus());
             workProjectFileList.add(workProjectFile);
             Integer useCount = file.getUseCount();
             file.setUseCount(useCount + 1);
         }
         workProjectFileMapper.insertList(workProjectFileList);
         for (File file : fileList) {
+            fileMapper.updateByPrimaryKey(file);
+        }
+        //2020-01-05增加参考方案的部分
+        List<WorkProjectFile> workProjectReferenceSolutionList = new ArrayList<>();
+        for (File file : referenceSolutionList) {
+            WorkProjectFile workProjectFile = buildWorkProjectFile(file, workProject);
+            //设置文件的使用类型为参考方案
+            workProjectFile.setFileType(Constants.File.FileTypes.REFERENCE_SOLUTION.getStatus());
+            workProjectReferenceSolutionList.add(workProjectFile);
+            Integer useCount = file.getUseCount();
+            file.setUseCount(useCount + 1);
+        }
+        workProjectFileMapper.insertList(workProjectReferenceSolutionList);
+        for (File file : referenceSolutionList) {
             fileMapper.updateByPrimaryKey(file);
         }
         return Result.success();
@@ -117,8 +138,14 @@ public class WorkProjectService {
         List<WorkProjectListDTO> workProjectListDTOList = workProjectMapper.getWorkProjectList();
         PageInfo<WorkProjectListDTO> pageInfo = new PageInfo(workProjectListDTOList);
         for (WorkProjectListDTO workProjectListDTO : workProjectListDTOList) {
-            List<String> fileNameList = fileMapper.getFileNameByWorkProjectId(workProjectListDTO.getId());
+            //2020-01-05 区分资料和参考方案，更换mapper方法，增加文件类型的控制
+            // List<String> fileNameList = fileMapper.getFileNameByWorkProjectId(workProjectListDTO.getId());
+            // 获取文件名称list
+            List<String> fileNameList = fileMapper.getFileNameByWorkProjectIdAndType(workProjectListDTO.getId(),Constants.File.FileTypes.INFORMATION.getStatus());
             workProjectListDTO.setFileName(fileNameList);
+            // 获取参考方案名称list
+            List<String>  referenceSolutionNameList = fileMapper.getFileNameByWorkProjectIdAndType(workProjectListDTO.getId(),Constants.File.FileTypes.REFERENCE_SOLUTION.getStatus());
+            workProjectListDTO.setReferenceSolutionName(referenceSolutionNameList);
         }
         BaseListDTO<WorkProjectListDTO> baseListDTO = new BaseListDTO<>(pageInfo.getTotal(), pageInfo.getList());
         return Result.success(baseListDTO);
@@ -136,8 +163,15 @@ public class WorkProjectService {
         if (null == workProjectInfoDTO) {
             throw new BaseException(Error.WORK_PROJECT_IS_NOT_EXIST);
         }
-        List<FileListDTO> fileListDTOList = fileMapper.getFileListDTOByWorkProjectId(workProjectInfoDTO.getId());
+        //2020-01-05 区分资料和参考方案，更换mapper方法，增加文件类型的控制
+        //List<FileListDTO> fileListDTOList = fileMapper.getFileListDTOByWorkProjectId(workProjectInfoDTO.getId());
+        //获取资料文件
+        List<FileListDTO> fileListDTOList = fileMapper.getFileListDTOByWorkProjectIdAndType(workProjectInfoDTO.getId(),Constants.File.FileTypes.INFORMATION.getStatus());
         workProjectInfoDTO.setFilelist(fileListDTOList);
+        //获取参考方案文件
+        List<FileListDTO> referenceSolutionListDTOList = fileMapper.getFileListDTOByWorkProjectIdAndType(workProjectInfoDTO.getId(),Constants.File.FileTypes.REFERENCE_SOLUTION.getStatus());
+        workProjectInfoDTO.setReferenceSolutionList(referenceSolutionListDTOList);
+
         return Result.success(workProjectInfoDTO);
     }
 
@@ -163,6 +197,11 @@ public class WorkProjectService {
         if (CollectionUtils.isEmpty(fileList) || param.getFileIdList().size() != fileList.size()) {
             throw new BaseException(Error.WORK_PROJECT_ASSOCIATE_FILE_HAS_NOT_EXIST);
         }
+        //2020-01-05增加参考方案的部分，这里参考方案是否可以为空
+        List<File> referenceSolutionList = fileManager.getFileByIdList(param.getReferenceSolutionList());
+        if (CollectionUtils.isEmpty(referenceSolutionList) || param.getReferenceSolutionList().size() != referenceSolutionList.size()) {
+            throw new BaseException(Error.WORK_PROJECT_ASSOCIATE_FILE_HAS_NOT_EXIST);
+        }
         workProjectFileMapper.deleteByWorkProjectId(workProject.getId());
         workProject.setUpdatetime(new Date());
         workProject.setUpdaterId(param.getLoginTeacherId());
@@ -172,12 +211,28 @@ public class WorkProjectService {
         List<WorkProjectFile> workProjectFileList = new ArrayList<>();
         for (File file : fileList) {
             WorkProjectFile workProjectFile = buildWorkProjectFile(file, workProject);
+            //设置文件的使用类型为资料
+            workProjectFile.setFileType(Constants.File.FileTypes.INFORMATION.getStatus());
             workProjectFileList.add(workProjectFile);
             Integer useCount = file.getUseCount();
             file.setUseCount(useCount + 1);
         }
         workProjectFileMapper.insertList(workProjectFileList);
         for (File file : fileList) {
+            fileMapper.updateByPrimaryKey(file);
+        }
+        //2020-01-05增加参考方案的部分
+        List<WorkProjectFile> workProjectReferenceSolutionList = new ArrayList<>();
+        for (File file : referenceSolutionList) {
+            WorkProjectFile workProjectFile = buildWorkProjectFile(file, workProject);
+            //设置文件的使用类型为参考方案
+            workProjectFile.setFileType(Constants.File.FileTypes.REFERENCE_SOLUTION.getStatus());
+            workProjectReferenceSolutionList.add(workProjectFile);
+            Integer useCount = file.getUseCount();
+            file.setUseCount(useCount + 1);
+        }
+        workProjectFileMapper.insertList(workProjectReferenceSolutionList);
+        for (File file : referenceSolutionList) {
             fileMapper.updateByPrimaryKey(file);
         }
 
@@ -212,8 +267,14 @@ public class WorkProjectService {
     public Result getAllWorkProjectList() {
         List<WorkProjectListDTO> workProjectListDTOList = workProjectMapper.getWorkProjectList();
         for (WorkProjectListDTO workProjectListDTO : workProjectListDTOList) {
-            List<String> fileNameList = fileMapper.getFileNameByWorkProjectId(workProjectListDTO.getId());
+            //2020-01-05 区分资料和参考方案，更换mapper方法，增加文件类型的控制
+            //List<String> fileNameList = fileMapper.getFileNameByWorkProjectId(workProjectListDTO.getId());
+            // 获取文件名称list
+            List<String> fileNameList = fileMapper.getFileNameByWorkProjectIdAndType(workProjectListDTO.getId(),Constants.File.FileTypes.INFORMATION.getStatus());
             workProjectListDTO.setFileName(fileNameList);
+            // 获取参考方案名称list
+            List<String>  referenceSolutionNameList = fileMapper.getFileNameByWorkProjectIdAndType(workProjectListDTO.getId(),Constants.File.FileTypes.REFERENCE_SOLUTION.getStatus());
+            workProjectListDTO.setReferenceSolutionName(referenceSolutionNameList);
         }
         BaseListDTO<WorkProjectListDTO> baseListDTO = new BaseListDTO<>(workProjectListDTOList.size(), workProjectListDTOList);
         return Result.success(baseListDTO);
